Remove JMediaFileDao
