+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/service/PlaylistService.java
@@ -38,7 +38,7 @@ import java.util.concurrent.ExecutionException;
 import chameleon.playlist.SpecificPlaylist;
 import chameleon.playlist.SpecificPlaylistFactory;
 import chameleon.playlist.SpecificPlaylistProvider;
-import com.tesshu.jpsonic.dao.JMediaFileDao;
+import com.tesshu.jpsonic.dao.MediaFileDao;
 import com.tesshu.jpsonic.dao.PlaylistDao;
 import com.tesshu.jpsonic.domain.JpsonicComparators;
 import com.tesshu.jpsonic.domain.MediaFile;
@@ -71,7 +71,7 @@ public class PlaylistService {
 
     private static final String EXPORT_PLAYLIST_FORMAT = "m3u";
 
-    private final JMediaFileDao mediaFileDao;
+    private final MediaFileDao mediaFileDao;
     private final PlaylistDao playlistDao;
     private final SecurityService securityService;
     private final SettingsService settingsService;
@@ -79,7 +79,7 @@ public class PlaylistService {
     private final List<PlaylistImportHandler> importHandlers;
     private final JpsonicComparators comparators;
 
-    public PlaylistService(JMediaFileDao mediaFileDao, PlaylistDao playlistDao, SecurityService securityService,
+    public PlaylistService(MediaFileDao mediaFileDao, PlaylistDao playlistDao, SecurityService securityService,
             SettingsService settingsService, List<PlaylistExportHandler> exportHandlers,
             List<PlaylistImportHandler> importHandlers, JpsonicComparators comparators) {
         this.mediaFileDao = mediaFileDao;
